+++ K8sUtils/private/Get-CurrentTime.ps1
@@ -1,10 +1,10 @@
 # if not UTC allow setting hours offset.
-$script:UtcOffset = [TimeSpan]::Zero
+$script:UtcOffset = [DateTimeOffset]::Now.Offset
 
 # helper to get the current time as configured
 function Get-CurrentTime {
     param(
         [TimeSpan] $Offset = [TimeSpan]::Zero
     )
-    return (Get-Date).Add($script:UtcOffset).Add($Offset)
+    return (Get-Date).Subtract($script:UtcOffset).Add($Offset)
 }